+++ src/snippets/style-tags.liquid
@@ -1,30 +1,32 @@
-{% for stylesheet in styles %}
-  {% if stylesheet.name == 'theme' -%}
-    <link rel="stylesheet" type="text/css" href="{% raw %}{{- 'theme.css' | asset_url -}}{% endraw %}" />
-  {% else -%}
-    {%- raw -%}
-    {% capture stylesheet_name %}{% endraw %}{{ stylesheet.name }}{% raw %}{% endcapture %}
-    {% if template == stylesheet_name %}
-    {%- endraw -%}
-      {% if stylesheet.critical %}
-      {%- assign template_critical_styles = 'template.critical.' | append: stylesheet.name | append: '.css' -%}
-      {%- assign template_styles = 'template.' | append: stylesheet.name | append: '.css' -%}
+{%- comment -%} EDIT WITH CAUTION {%- endcomment -%}
+{%- if styles.size > 0 -%}
+  <!-- Style Tags -->
+  {% for stylesheet in styles -%}
+    {% if stylesheet.name == 'theme' -%}
+      <link rel="stylesheet" type="text/css" href="{% raw %}{{- 'theme.css' | asset_url -}}{% endraw %}" />
+    {% else -%}
       {%- raw -%}
-      {% capture template_critical_stylesheet %}{% endraw %}{{ template_critical_styles }}{% raw %}{% endcapture %}
-      {% capture template_stylesheet %}{% endraw %}{{ template_styles }}{% raw %}{% endcapture %}
-      {%- endraw -%}
-      <link rel="stylesheet" type="text/css" href="{% raw %}{{- template_critical_stylesheet | asset_url -}}{% endraw %}" />
-      <link rel="preload" href="{% raw %}{{- template_stylesheet | asset_url -}}{% endraw %}" as="style" onload="this.onload=null;this.rel='stylesheet'" />
-      {% else %}
-      {%- assign template_styles = 'template.' | append: stylesheet.name | append: '.css' -%}
+      {%- capture stylesheet_template_name -%}{% endraw %}{{ stylesheet.name }}{% raw %}{%- endcapture -%}
+      {%- if template == stylesheet_template_name -%}{%- endraw -%}
+        {% if stylesheet.critical %}
+        {%- assign template_critical_styles = 'template.critical.' | append: stylesheet.name | append: '.css' -%}
+        {%- assign template_styles = 'template.' | append: stylesheet.name | append: '.css' -%}
+        {%- raw -%}
+        {%- capture template_critical_stylesheet -%}{% endraw %}{{ template_critical_styles }}{% raw %}{%- endcapture -%}
+        {%- capture template_stylesheet -%}{% endraw %}{{ template_styles }}{% raw %}{%- endcapture -%}
+        {%- endraw -%}
+        <link rel="stylesheet" type="text/css" href="{% raw %}{{- template_critical_stylesheet | asset_url -}}{% endraw %}" />
+        <link rel="preload" href="{% raw %}{{- template_stylesheet | asset_url -}}{% endraw %}" as="style" onload="this.onload=null;this.rel='stylesheet'" />
+        {%- else %}
+        {%- assign template_styles = 'template.' | append: stylesheet.name | append: '.css' -%}
+        {%- raw -%}
+        {%- capture template_stylesheet -%}{% endraw %}{{ template_styles }}{% raw %}{%- endcapture -%}
+        {%- endraw -%}
+        <link rel="stylesheet" type="text/css"  href="{% raw %}{{- template_stylesheet | asset_url -}}{% endraw %}" />
+        {%- endif %}
       {%- raw -%}
-      {% capture template_stylesheet %}{% endraw %}{{ template_styles }}{% raw %}{% endcapture %}
+      {%- endif -%}
       {%- endraw -%}
-
-      <link rel="stylesheet" type="text/css"  href="{% raw %}{{- template_stylesheet | asset_url -}}{% endraw %}" />
-      {% endif %}
-    {%- raw -%}
-    {% endif %}
-    {%- endraw -%}
-  {% endif -%}
-{% endfor %}
+    {% endif -%}
+  {% endfor %}
+{%- endif -%}
